refactor

+++ src/main/java/dev/zux13/currency/exchange/entity/ExchangeRate.java
@@ -4,6 +4,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
@@ -13,5 +15,5 @@ public class ExchangeRate {
     private Long id;
     private Long baseCurrencyId;
     private Long targetCurrencyId;
-    private double rate;
+    private BigDecimal rate;
 }
